Add addressing mode 2 decoding (load and store instructions)
This is a fairly naive implementation and will probably have to go
through a couple of optimisation revisions. I'm sure there are tricks to
make it faster... The behaviour of the function is tested by a few unit
tests but not all cases are tested for specifically. For example, the
offset addition and substraction are not tested for all 9 offset formats,
however, the tests, that do test it, imply that it should be working
correctly for all formats. More should be added as the function becomes
more complex.